Apps/documents/requirements.txt
elasticsearch-dsl>=7.4.0
python-magic>=0.4.27
Pillow>=9.0.0
django-storages>=1.13.2
boto3>=1.26.0  # For S3 storage support
python-dotenv>=0.19.0 